arbitrage_app/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
prometheus-client==0.19.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
